+++ server/app.js
@@ -42,6 +42,7 @@ const config     = require('./config')
     , Servers    = require('./core/http')
     , Routes     = require('./core/routes')
     , Watcher    = require('./core/watcher')
+    , Processor  = require('./core/process')
     , FilterIP   = require('./middlewares/ip/main')
     , KillOutSe  = require('./middlewares/session/kill')
     , I18N       = require('./libs/i18n/i18n')
@@ -216,6 +217,7 @@ function App() {
   // Start the HTTP & HTTPS servers:
   Routes.start(app, i18n, dbi, dbn);
   Watcher.start(app, i18n, dbi, dbn);
+  Processor.start(app, i18n, dbi, dbn);
   Servers.startHttp(app);
   Servers.startHttps(app, __dirname);
 
@@ -224,6 +226,7 @@ function App() {
   //   // Start the HTTP & HTTPS servers:
   //   Routes.start(app, i18n, dbi, dbn, dbmo);
   //   Watcher.start(app, i18n, dbi, dbn, dbmo);
+  //   Processor.start(app, i18n, dbi, dbn, dbmo);
   //   Servers.startHttp(app);
   //   Servers.startHttps(app, __dirname);
   // });